Removed unnecessary memcpy in calc_chunk
This was I believe the main idea behind pull requests #6 and #8. While
calc_chunk does indeed have an unnecessary memcpy, removing is it quite
easy: just make the use of the input calculation buffer optional, and
return a pointer instead of a bool.

Thanks to jan2000 for pointing out the unnecessary memcpy, but I do not
think it justifies removing calc_chunk, which brings structure clarity
to the design.

With this commit, I get pretty much the same performance improvement as
with pull request #6.