Implemented a streaming API
Since this is a generalization of the original API, the original API
(function,  really) is reimplemented in terms of the new one. And since
the original implementation was not structured around this
generalization from the beginning, this commit is pretty much a complete
rewrite of the implementation.

Special attention was given to performance, or more specifically
ensuring that the performance is not worsened by this commit for the
original case. While there is no extensive proof of that, running the
original tests, which themselves are quite extensive, on the developer's
machine, shows no measurable performance penalty,

Interestingly, the use of `inline` for the function that makes the
calculations on a chunk of data has a measurable impact on performance
(a systematic one to two percent impact could be measured). For the
record, that difference could be measured with gcc 10.3 on a x86-64
machine.